+++ src/paas_charm/charm.py
@@ -11,7 +11,6 @@ import typing
 import ops
 from charms.data_platform_libs.v0.data_interfaces import DatabaseRequiresEvent
 from charms.openfga_k8s.v1.openfga import OpenFGARequires
-from charms.redis_k8s.v0.redis import RedisRelationCharmEvents
 from charms.smtp_integrator.v0.smtp import SmtpRequires
 from charms.traefik_k8s.v2.ingress import IngressPerAppRequirer
 from ops import RelationMeta
@@ -34,7 +33,6 @@ from paas_charm.paas_config import (
 )
 from paas_charm.peers import Peers
 from paas_charm.rabbitmq import RabbitMQRequires
-from paas_charm.redis import PaaSRedisRequires
 from paas_charm.s3 import PaaSS3Requirer
 from paas_charm.saml import PaaSSAMLRequirer
 from paas_charm.secret_key import SecretKeyStorage
@@ -54,7 +52,6 @@ class PaasCharm(abc.ABC, ops.CharmBase):  # pylint: disable=too-many-instance-at
     """PaasCharm base charm service mixin.
 
     Attrs:
-        on: charm events replaced by Redis ones for the Redis charm library.
         framework_config_class: base class for the framework config.
         paas_config_framework_fields: Mapping from framework fields to paas-config fields.
     """
@@ -78,8 +75,6 @@ class PaasCharm(abc.ABC, ops.CharmBase):  # pylint: disable=too-many-instance-at
         # of the application. State only supposed to live within the lifecycle of the container.
         return pathlib.Path(f"/tmp/{self._framework_name}/state")  # nosec: B108
 
-    on = RedisRelationCharmEvents()
-
     def __init__(self, framework: ops.Framework, framework_name: str) -> None:
         """Initialize the instance.
 
@@ -100,7 +95,6 @@ class PaasCharm(abc.ABC, ops.CharmBase):  # pylint: disable=too-many-instance-at
         self._database_requirers = make_database_requirers(self, self.app.name)
 
         requires: dict[str, RelationMeta] = self.framework.meta.requires
-        self._redis = self._init_redis(requires)
         self._valkey = self._init_valkey(requires)
         self._s3 = self._init_s3(requires)
         self._saml = self._init_saml(requires)
@@ -178,30 +172,6 @@ class PaasCharm(abc.ABC, ops.CharmBase):  # pylint: disable=too-many-instance-at
             self.on[self._workload_config.container_name].pebble_ready,
             self._reconcile_without_migrations,
         )
-
-    def _init_redis(self, requires: dict[str, RelationMeta]) -> "PaaSRedisRequires | None":
-        """Initialize the Redis relation if its required.
-
-        Args:
-            requires: relation requires dictionary from metadata
-
-        Returns:
-            Returns the Redis relation or None
-        """
-        _redis = None
-        if "redis" in requires and requires["redis"].interface_name == "redis":
-            try:
-                _redis = PaaSRedisRequires(charm=self, relation_name="redis")
-                self.framework.observe(
-                    self.on.redis_relation_updated, self._reconcile_with_migrations
-                )
-            except NameError:
-                logger.exception(
-                    "Missing charm library,                               "
-                    "please run `charmcraft fetch-lib charms.redis_k8s.v0.redis`"
-                )
-
-        return _redis
 
     def _init_valkey(self, requires: dict[str, RelationMeta]) -> "ValkeyClientRequirer | None":
         """Initialize the Valkey relation if it is required.
@@ -619,10 +589,6 @@ class PaasCharm(abc.ABC, ops.CharmBase):  # pylint: disable=too-many-instance-at
             requires: relation requires dictionary from metadata
             charm_state: current charm state
         """
-        if self._redis and not charm_state.integrations.redis:
-            if not requires["redis"].optional:
-                yield "redis"
-
         if self._s3 and not charm_state.integrations.s3:
             if not requires["s3"].optional:
                 yield "s3"
@@ -640,22 +606,30 @@ class PaasCharm(abc.ABC, ops.CharmBase):  # pylint: disable=too-many-instance-at
             requires: relation requires dictionary from metadata
             charm_state: current charm state
         """
-        if self._saml and not charm_state.integrations.saml:
-            if not requires["saml"].optional:
-                yield "saml"
+        if self._saml and not charm_state.integrations.saml and not requires["saml"].optional:
+            yield "saml"
 
-        if self._tracing and not charm_state.integrations.tracing:
-            if not requires["tracing"].optional:
-                yield "tracing"
+        if (
+            self._tracing
+            and not charm_state.integrations.tracing
+            and not requires["tracing"].optional
+        ):
+            yield "tracing"
 
-        if self._smtp and not charm_state.integrations.smtp:
-            if not requires["smtp"].optional:
-                yield "smtp"
+        if self._smtp and not charm_state.integrations.smtp and not requires["smtp"].optional:
+            yield "smtp"
 
         if self._oauth and not charm_state.integrations.oauth:
             oauth_endpoint_name = get_endpoints_by_interface_name(requires, "oauth")[0][0]
             if not requires[oauth_endpoint_name].optional:
                 yield "oauth"
+
+        if (
+            self._valkey
+            and not charm_state.integrations.valkey
+            and not requires["valkey"].optional
+        ):
+            yield "valkey"
 
     def _missing_required_integrations(
         self, charm_state: CharmState
@@ -738,7 +712,6 @@ class PaasCharm(abc.ABC, ops.CharmBase):  # pylint: disable=too-many-instance-at
             peers=self._peers,
             integration_requirers=IntegrationRequirers(
                 databases=self._database_requirers,
-                redis=self._redis,
                 valkey=self._valkey,
                 rabbitmq=self._rabbitmq,
                 s3=self._s3,