+++ tests/integration/integrations/test_valkey.py
@@ -15,11 +15,11 @@ from tests.integration.types import App
 logger = logging.getLogger(__name__)
 
 
-@pytest.mark.skip(reason="Valkey is unstable rn")
 @pytest.mark.parametrize(
     "app_fixture, port, endpoint",
     [
         ("flask_app", 8000, "valkey/status"),
+        ("spring_boot_app", 8080, "valkey/status"),
     ],
 )
 def test_with_valkey(